# MortgageCalculator/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mortgage Calculator API",
        "version": "1.0.0"
    },
    "paths": {
        "/authentication/register": {
            "post": {
                "tags": [
                    "/authentication"
                ],
                "summary": "sign up",
                "description": "register by email, password",
                "operationId": "75462ced126f8e0caa3c79d9a0d3ded0",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "John Johnson"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful user registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The user has been successfully registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Wrong credentials for registration"
                    }
                }
            }
        },
        "/authentication/login": {
            "post": {
                "tags": [
                    "/authentication"
                ],
                "summary": "sign in",
                "description": "login by email, password",
                "operationId": "65c1a2df4b8dab8aa01a48a5251e395f",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                }
            }
        },
        "/mortgage-programs": {
            "get": {
                "tags": [
                    "/mortgage-programs"
                ],
                "operationId": "97c965a8219383e58fd4d36460d98ce9",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/mortgage-programs"
                ],
                "operationId": "d9e3cb8c5f67b1d1c3195a986f1e55be",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_name",
                                    "program_name",
                                    "realty_type",
                                    "interest_rate",
                                    "minimum_payment",
                                    "maximum_amoumt"
                                ],
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "format": "bank name",
                                        "example": "Sberbank"
                                    },
                                    "program_name": {
                                        "type": "string",
                                        "format": "program name",
                                        "example": "невыгодная ипотека"
                                    },
                                    "realty_type": {
                                        "type": "string",
                                        "format": "realty type",
                                        "example": "жилое"
                                    },
                                    "interest_rate": {
                                        "type": "decimal",
                                        "format": "interest rate",
                                        "example": "10"
                                    },
                                    "minimum_payment": {
                                        "type": "decimal",
                                        "format": "minimum payment",
                                        "example": "100000"
                                    },
                                    "maximum_amount": {
                                        "type": "decimal",
                                        "format": "maximum amount",
                                        "example": "500000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mortgage-programs/{id}": {
            "get": {
                "tags": [
                    "/mortgage-programs"
                ],
                "operationId": "1fe5194d7df099a74c9863bf86eafbb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/mortgage-programs"
                ],
                "operationId": "c52eb1aaad6d696d4811020a6f1bb7f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_name",
                                    "program_name",
                                    "realty_type",
                                    "interest_rate",
                                    "minimum_payment",
                                    "maximum_amoumt"
                                ],
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "format": "bank name",
                                        "example": "Sberbank"
                                    },
                                    "program_name": {
                                        "type": "string",
                                        "format": "program name",
                                        "example": "невыгодная ипотека"
                                    },
                                    "realty_type": {
                                        "type": "string",
                                        "format": "realty type",
                                        "example": "жилое"
                                    },
                                    "interest_rate": {
                                        "type": "decimal",
                                        "format": "interest rate",
                                        "example": "10"
                                    },
                                    "minimum_payment": {
                                        "type": "decimal",
                                        "format": "minimum payment",
                                        "example": "100000"
                                    },
                                    "maximum_amount": {
                                        "type": "decimal",
                                        "format": "maximum amount",
                                        "example": "500000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/mortgage-programs"
                ],
                "operationId": "9be44a941ef2c29e069692890dda960c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}